requirements.txt
Pillow==8.2.0
requests==2.25.1
tk==0.1.0
